+++ src/ui/menu/main/signup_popup.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <defs.hpp>
 
-class GlobedSignupPopup : public geode::Popup<>, public CommentUploadDelegate {
+class GlobedSignupPopup : public geode::Popup<>, public UploadMessageDelegate {
 public:
     constexpr static float POPUP_WIDTH = 180.f;
     constexpr static float POPUP_HEIGHT = 80.f;
@@ -12,7 +12,7 @@ protected:
     cocos2d::CCLabelBMFont* statusMessage;
 
     std::string storedAuthcode;
-    int storedLevelId;
+    int storedAccountId;
 
     bool setup() override;
     void keyDown(cocos2d::enumKeyCodes key) override;
@@ -20,12 +20,11 @@ protected:
     void onFailure(const std::string_view message);
     void onSuccess();
 
-    void onChallengeCreated(int levelId, const std::string_view chtoken);
+    void onChallengeCreated(int accountId, const std::string_view chtoken);
     void onChallengeCompleted(const std::string_view authcode);
 
-    void commentUploadFinished(int) override;
-    void commentUploadFailed(int, CommentError) override;
-    void commentDeleteFailed(int, int) override;
+    void uploadMessageFinished(int) override;
+    void uploadMessageFailed(int) override;
 
     void onDelayedChallengeCompleted();
 };